(PC-18448)[PRO] test: add test to check initial value
