+++ pro/src/app/AppRouter/subroutes_offerindividualwizard_map.tsx
@@ -7,16 +7,6 @@ import { Offer } from 'pages/OfferIndividualWizard/Offer'
 import { PriceCategories } from 'pages/OfferIndividualWizard/PriceCategories'
 import { Stocks } from 'pages/OfferIndividualWizard/Stocks'
 import { Summary } from 'pages/OfferIndividualWizard/Summary'
-import SignupConfirmation from 'pages/Signup/SignupConfirmation/SignupConfirmation'
-import SignupContainer from 'pages/Signup/SignupContainer/SignupContainer'
-import SignUpValidation from 'pages/Signup/SignUpValidation'
-import { Activity } from 'screens/SignupJourneyForm/Activity'
-import { OffererAuthentication } from 'screens/SignupJourneyForm/Authentication'
-import { ConfirmedAttachment } from 'screens/SignupJourneyForm/ConfirmedAttachment'
-import { Offerer } from 'screens/SignupJourneyForm/Offerer'
-import { Offerers as SignupJourneyOfferers } from 'screens/SignupJourneyForm/Offerers'
-import { Validation } from 'screens/SignupJourneyForm/Validation'
-import { Welcome } from 'screens/SignupJourneyForm/Welcome'
 
 import type { IRoute, RouteDefinition } from './routes_map'
 
@@ -113,82 +103,7 @@ export const routesOfferIndividualWizard: IRoute[] = [
   },
 ]
 
-export const routesSignup: IRoute[] = [
-  {
-    element: <SignupContainer />,
-    parentPath: '/inscription',
-    path: '/',
-    title: 'S’inscrire',
-  },
-  {
-    element: <SignupConfirmation />,
-    parentPath: '/inscription',
-    path: '/confirmation',
-    title: 'S’inscrire',
-  },
-  {
-    element: <SignUpValidation />,
-    parentPath: '/inscription',
-    path: '/validation/:token',
-    title: 'S’inscrire',
-  },
-]
-
-export const routesSignupJourney: IRoute[] = [
-  {
-    element: <Welcome />,
-    parentPath: '/parcours-inscription',
-    path: '/',
-    title: '',
-  },
-  {
-    element: <Offerer />,
-    parentPath: '/parcours-inscription',
-    path: '/structure',
-    title: '',
-  },
-  {
-    element: <SignupJourneyOfferers />,
-    parentPath: '/parcours-inscription',
-    path: '/structure/rattachement',
-    title: '',
-  },
-  {
-    element: <ConfirmedAttachment />,
-    parentPath: '/parcours-inscription',
-    path: '/structure/rattachement/confirmation',
-    title: '',
-  },
-  {
-    element: <OffererAuthentication />,
-    parentPath: '/parcours-inscription',
-    path: '/authentification',
-    title: '',
-  },
-  {
-    element: <OffererAuthentication />,
-    parentPath: '/parcours-inscription',
-    path: '/authentification',
-    title: '',
-  },
-  {
-    element: <Activity />,
-    parentPath: '/parcours-inscription',
-    path: '/activite',
-    title: '',
-  },
-  {
-    element: <Validation />,
-    parentPath: '/parcours-inscription',
-    path: '/validation',
-    title: '',
-  },
-]
-
-export const subroutesDefinitions: RouteDefinition[] = [
-  ...routesOfferIndividualWizard,
-  ...routesSignup,
-  ...routesSignupJourney,
-].map(({ path, parentPath, title }) => {
-  return { path, parentPath, title }
-})
+export const routesOfferIndividualWizardDefinitions: RouteDefinition[] =
+  routesOfferIndividualWizard.map(({ path, parentPath, title }) => {
+    return { path, parentPath, title }
+  })
