+++ pro/src/app/AppRouter/subroutes_signup_map.tsx
@@ -1,0 +1,36 @@
+/* No need to test this file */
+/* istanbul ignore file */
+import React from 'react'
+
+import SignupConfirmation from 'pages/Signup/SignupConfirmation/SignupConfirmation'
+import SignupContainer from 'pages/Signup/SignupContainer/SignupContainer'
+import SignUpValidation from 'pages/Signup/SignUpValidation'
+
+import type { IRoute, RouteDefinition } from './routes_map'
+
+export const routesSignup: IRoute[] = [
+  {
+    element: <SignupContainer />,
+    parentPath: '/inscription',
+    path: '/',
+    title: 'S’inscrire',
+  },
+  {
+    element: <SignupConfirmation />,
+    parentPath: '/inscription',
+    path: '/confirmation',
+    title: 'S’inscrire',
+  },
+  {
+    element: <SignUpValidation />,
+    parentPath: '/inscription',
+    path: '/validation/:token',
+    title: 'S’inscrire',
+  },
+]
+
+export const routesSignupDefinitions: RouteDefinition[] = routesSignup.map(
+  ({ path, parentPath, title }) => {
+    return { path, parentPath, title }
+  }
+)
