+++ pro/src/hooks/usePageTitle.ts
@@ -3,7 +3,9 @@ import { useEffect } from 'react'
 import { matchPath, useLocation } from 'react-router-dom'
 
 import { routesDefinitions, RouteDefinition } from 'app/AppRouter/routes_map'
-import { subroutesDefinitions } from 'app/AppRouter/subroutes_map'
+import { routesOfferIndividualWizardDefinitions } from 'app/AppRouter/subroutes_offerindividualwizard_map'
+import { routesSignupDefinitions } from 'app/AppRouter/subroutes_signup_map'
+import { routesSignupJourneyDefinitions } from 'app/AppRouter/subroutes_signupjourney_map'
 
 const usePageTitle = (): LocationListener | void => {
   const location = useLocation()
@@ -11,7 +13,9 @@ const usePageTitle = (): LocationListener | void => {
   useEffect(() => {
     const currentRoute = [
       ...routesDefinitions,
-      ...subroutesDefinitions,
+      ...routesOfferIndividualWizardDefinitions,
+      ...routesSignupDefinitions,
+      ...routesSignupJourneyDefinitions,
       // @ts-expect-error Property findLast does not exist on RouteDefinition[]
     ].findLast(
       ({ path, parentPath }: RouteDefinition) =>
