+++ pro/src/pages/OfferIndividualWizard/OfferIndividualWizard.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useLocation, Route, Routes, useParams } from 'react-router-dom'
 
 import { IRoute } from 'app/AppRouter/routes_map'
-import { routesOfferIndividualWizard } from 'app/AppRouter/subroutes_map'
+import { routesOfferIndividualWizard } from 'app/AppRouter/subroutes_offerindividualwizard_map'
 import { OfferIndividualContextProvider } from 'context/OfferIndividualContext'
 import useCurrentUser from 'hooks/useCurrentUser'
 import { parse } from 'utils/query-string'
